[#61] refactor: jwt/userRepository 패키지 common 모듈로 이동
