+++ b2c/src/main/java/itcast/auth/dto/response/AccessTokenResponse.java
@@ -2,9 +2,6 @@ package itcast.auth.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import lombok.Getter;
-import lombok.Setter;
-
 public record AccessTokenResponse(
         @JsonProperty(value = "access_token", required = true)
         String accessToken
