+++ b2c/src/main/java/itcast/auth/dto/response/KakaoUserInfo.java
@@ -2,5 +2,4 @@ package itcast.auth.dto.response;
 
 public record KakaoUserInfo(
         String kakaoEmail
-) {
-}
+) {}
