+++ b2c/src/main/java/itcast/user/controller/UserController.java
@@ -3,15 +3,14 @@ package itcast.user.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import itcast.ResponseTemplate;
-import itcast.auth.jwt.CheckAuth;
-import itcast.auth.jwt.LoginMember;
+import itcast.jwt.CheckAuth;
+import itcast.jwt.LoginMember;
 import itcast.user.application.UserService;
 import itcast.user.dto.request.ProfileCreateRequest;
 import itcast.user.dto.request.ProfileUpdateRequest;
