+++ common/src/main/java/itcast/ResponseTemplate.java
@@ -1,9 +1,6 @@
 package itcast;
 
 import lombok.Getter;
-import lombok.Setter;
-
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 
 @Getter
