+++ common/src/main/java/itcast/jwt/CheckAuth.java
@@ -1,11 +1,11 @@
-package itcast.auth.jwt;
+package itcast.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.PARAMETER)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LoginMember {
+public @interface CheckAuth {
 }
