+++ common/src/main/java/itcast/jwt/LoginMember.java
@@ -1,11 +1,11 @@
-package itcast.auth.jwt;
+package itcast.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CheckAuth {
+public @interface LoginMember {
 }
